+++ myBenchmark/myBenchmark.FSharp/myBenchmark.FSharp.fsproj
@@ -5,7 +5,7 @@
   </PropertyGroup>
 
   <ItemGroup>
-    <Compile Include="Library.fs" />
+    <Compile Include="FizzBuzz.fs" />
   </ItemGroup>
 
 </Project>
